Consolidation + SidePanelView Icons
Moved a decent chunk of style and resource code into the App.axaml file so it can be used globally. Also implemented the download icon/animation/functionality into the sidepanelview. This all works now but still needs to be polished as the margins slightly change and the windows slightly move on button press which is annoying.