Added ViewModels + Views
Added new ViewModels and Views for Header and Control Panel. Added these both to use the MainWindow DataContext so now all panels can communicate with each other.

Also added some placeholders for the search bar and other buttons as well as a new `View` menu in the File Menu at the top.