Changes to auto updater
Tested the auto updater on Linux, OSX and Windows. Does not work properly on Linux and OSX because of permissions issues (?) and its a real headache to try and fix. The auto updater works with no issues on Windows though. I've implemented a "Manual Update" button that simply takes the user to the latest release page on github so they can download it manually if using Linux or OSX. Will maybe revisit this in future.