First pass fix Linux binary analysis
First go at fixing Linux support. Previously linux binary could not be found due to misnamed fileName (still not sure if what I added is the correct name or not).

Added a separate function to try and determine the ProductVersion on Linux using exiftool. This is all untested on linux.